+++ com/baidu/android/imsdk/chatmessage/request/IMMediaDeleteSessionRequest.java
@@ -10,13 +10,20 @@ import com.baidu.android.imsdk.internal.Constants;
 import com.baidu.android.imsdk.internal.ListenerManager;
 import com.baidu.android.imsdk.utils.LogUtils;
 import com.baidu.android.imsdk.utils.Utility;
+import com.baidu.titan.sdk.runtime.FieldHolder;
+import com.baidu.titan.sdk.runtime.InitContext;
+import com.baidu.titan.sdk.runtime.InterceptResult;
+import com.baidu.titan.sdk.runtime.Interceptable;
+import com.baidu.titan.sdk.runtime.TitanRuntime;
 import java.security.NoSuchAlgorithmException;
 import java.util.Map;
 import org.json.JSONException;
 import org.json.JSONObject;
 /* loaded from: classes.dex */
 public class IMMediaDeleteSessionRequest extends IMMediaBaseHttpRequest {
+    public static /* synthetic */ Interceptable $ic = null;
     public static final String TAG = "IMMediaDeleteSessionRequest";
+    public transient /* synthetic */ FieldHolder $fh;
     public long mContacter;
     public long mContactorPauid;
     public String mContactorThirdid;
@@ -25,6 +32,20 @@ public class IMMediaDeleteSessionRequest extends IMMediaBaseHttpRequest {
     public long mLastTime;
 
     public IMMediaDeleteSessionRequest(Context context, long j, long j2, String str) {
+        Interceptable interceptable = $ic;
+        if (interceptable != null) {
+            InitContext newInitContext = TitanRuntime.newInitContext();
+            newInitContext.initArgs = r2;
+            Object[] objArr = {context, Long.valueOf(j), Long.valueOf(j2), str};
+            interceptable.invokeUnInit(65537, newInitContext);
+            int i2 = newInitContext.flag;
+            if ((i2 & 1) != 0) {
+                int i3 = i2 & 2;
+                newInitContext.thisArg = this;
+                interceptable.invokeInitBody(65537, newInitContext);
+                return;
+            }
+        }
         this.mContactorType = -1;
         this.mContactorPauid = -1L;
         this.mContext = context;
@@ -35,7 +56,9 @@ public class IMMediaDeleteSessionRequest extends IMMediaBaseHttpRequest {
 
     @Override // com.baidu.android.imsdk.utils.HttpHelper.Request
     public String getContentType() {
-        return "application/json";
+        InterceptResult invokeV;
+        Interceptable interceptable = $ic;
+        return (interceptable == null || (invokeV = interceptable.invokeV(1048576, this)) == null) ? "application/json" : (String) invokeV.objValue;
     }
 
     @Override // com.baidu.android.imsdk.chatmessage.request.IMMediaBaseHttpRequest, com.baidu.android.imsdk.utils.BaseHttpRequest, com.baidu.android.imsdk.utils.HttpHelper.Request
@@ -45,10 +68,15 @@ public class IMMediaDeleteSessionRequest extends IMMediaBaseHttpRequest {
 
     @Override // com.baidu.android.imsdk.utils.BaseHttpRequest, com.baidu.android.imsdk.utils.HttpHelper.Request
     public String getHost() {
-        if (getHostUrl() == null) {
-            return null;
+        InterceptResult invokeV;
+        Interceptable interceptable = $ic;
+        if (interceptable == null || (invokeV = interceptable.invokeV(Constants.METHOD_SEND_USER_MSG, this)) == null) {
+            if (getHostUrl() == null) {
+                return null;
+            }
+            return getHostUrl() + "rest/3.0/im/delete_session";
         }
-        return getHostUrl() + "rest/3.0/im/delete_session";
+        return (String) invokeV.objValue;
     }
 
     @Override // com.baidu.android.imsdk.chatmessage.request.IMMediaBaseHttpRequest, com.baidu.android.imsdk.utils.BaseHttpRequest, com.baidu.android.imsdk.utils.HttpHelper.Request
@@ -58,41 +86,49 @@ public class IMMediaDeleteSessionRequest extends IMMediaBaseHttpRequest {
 
     @Override // com.baidu.android.imsdk.utils.BaseHttpRequest, com.baidu.android.imsdk.utils.HttpHelper.Request
     public byte[] getRequestParameter() throws NoSuchAlgorithmException {
-        JSONObject jSONObject = new JSONObject();
-        try {
-            putCommonParams(jSONObject);
-            if (this.mContacter > 0) {
-                if (this.mContactorType == 2) {
-                    jSONObject.put("contacter_im_uk", this.mContacter);
-                    jSONObject.put("group_type", 3);
-                } else {
-                    jSONObject.put("contacter", Utility.transBDUID(String.valueOf(this.mContacter)));
+        InterceptResult invokeV;
+        Interceptable interceptable = $ic;
+        if (interceptable == null || (invokeV = interceptable.invokeV(1048580, this)) == null) {
+            JSONObject jSONObject = new JSONObject();
+            try {
+                putCommonParams(jSONObject);
+                if (this.mContacter > 0) {
+                    if (this.mContactorType == 2) {
+                        jSONObject.put("contacter_im_uk", this.mContacter);
+                        jSONObject.put("group_type", 3);
+                    } else {
+                        jSONObject.put("contacter", Utility.transBDUID(String.valueOf(this.mContacter)));
+                    }
                 }
+                if (this.mContactorType >= 0) {
+                    jSONObject.put("contacter_type", this.mContactorType);
+                }
+                if (this.mContactorPauid > 0) {
+                    jSONObject.put("contacter_pa_uid", this.mContactorPauid);
+                }
+                if (!TextUtils.isEmpty(this.mContactorThirdid)) {
+                    jSONObject.put("contacter_third_id", this.mContactorThirdid);
+                }
+                jSONObject.put("lastmsg_time", this.mLastTime);
+                jSONObject.put("sign", generateSign(jSONObject));
+            } catch (JSONException e2) {
+                LogUtils.e(TAG, "getRequestParameter Exception ", e2);
             }
-            if (this.mContactorType >= 0) {
-                jSONObject.put("contacter_type", this.mContactorType);
-            }
-            if (this.mContactorPauid > 0) {
-                jSONObject.put("contacter_pa_uid", this.mContactorPauid);
-            }
-            if (!TextUtils.isEmpty(this.mContactorThirdid)) {
-                jSONObject.put("contacter_third_id", this.mContactorThirdid);
-            }
-            jSONObject.put("lastmsg_time", this.mLastTime);
-            jSONObject.put("sign", generateSign(jSONObject));
-        } catch (JSONException e2) {
-            LogUtils.e(TAG, "getRequestParameter Exception ", e2);
+            return jSONObject.toString().getBytes();
         }
-        return jSONObject.toString().getBytes();
+        return (byte[]) invokeV.objValue;
     }
 
     @Override // com.baidu.android.imsdk.utils.BaseHttpRequest, com.baidu.android.imsdk.utils.HttpHelper.ResponseHandler
     public void onFailure(int i2, byte[] bArr, Throwable th) {
-        Pair<Integer, String> transErrorCode = transErrorCode(i2, bArr, th);
-        LogUtils.d(TAG, "onFailure error = " + transErrorCode.first + " errormsg = " + ((String) transErrorCode.second));
-        IMediaDeleteChatSessionListener iMediaDeleteChatSessionListener = (IMediaDeleteChatSessionListener) ListenerManager.getInstance().removeListener(this.mKey);
-        if (iMediaDeleteChatSessionListener != null) {
-            iMediaDeleteChatSessionListener.onMediaDeleteChatSessionResult(((Integer) transErrorCode.first).intValue(), (String) transErrorCode.second);
+        Interceptable interceptable = $ic;
+        if (interceptable == null || interceptable.invokeILL(1048581, this, i2, bArr, th) == null) {
+            Pair<Integer, String> transErrorCode = transErrorCode(i2, bArr, th);
+            LogUtils.d(TAG, "onFailure error = " + transErrorCode.first + " errormsg = " + ((String) transErrorCode.second));
+            IMediaDeleteChatSessionListener iMediaDeleteChatSessionListener = (IMediaDeleteChatSessionListener) ListenerManager.getInstance().removeListener(this.mKey);
+            if (iMediaDeleteChatSessionListener != null) {
+                iMediaDeleteChatSessionListener.onMediaDeleteChatSessionResult(((Integer) transErrorCode.first).intValue(), (String) transErrorCode.second);
+            }
         }
     }
 
@@ -100,23 +136,26 @@ public class IMMediaDeleteSessionRequest extends IMMediaBaseHttpRequest {
     public void onSuccess(int i2, byte[] bArr) {
         int i3;
         String str;
-        String str2 = new String(bArr);
-        LogUtils.d(TAG, "onSuccess resultContent = " + str2);
-        try {
-            JSONObject jSONObject = new JSONObject(str2);
-            i3 = jSONObject.optInt("error_code", 0);
-            str = jSONObject.optString("error_msg");
-            if (i3 == 0 && this.mContactorType == 2) {
-                ChatMessageDBManager.getInstance(this.mContext).deleteAllMsg(new ChatObject(this.mContext, 1, this.mContacter));
+        Interceptable interceptable = $ic;
+        if (interceptable == null || interceptable.invokeIL(1048582, this, i2, bArr) == null) {
+            String str2 = new String(bArr);
+            LogUtils.d(TAG, "onSuccess resultContent = " + str2);
+            try {
+                JSONObject jSONObject = new JSONObject(str2);
+                i3 = jSONObject.optInt("error_code", 0);
+                str = jSONObject.optString("error_msg");
+                if (i3 == 0 && this.mContactorType == 2) {
+                    ChatMessageDBManager.getInstance(this.mContext).deleteAllMsg(new ChatObject(this.mContext, 1, this.mContacter));
+                }
+            } catch (JSONException e2) {
+                LogUtils.e(TAG, "IMMediaDeleteSessionRequest JSONException", e2);
+                i3 = 1010;
+                str = Constants.ERROR_MSG_JSON_PARSE_EXCEPTION;
             }
-        } catch (JSONException e2) {
-            LogUtils.e(TAG, "IMMediaDeleteSessionRequest JSONException", e2);
-            i3 = 1010;
-            str = Constants.ERROR_MSG_JSON_PARSE_EXCEPTION;
-        }
-        IMediaDeleteChatSessionListener iMediaDeleteChatSessionListener = (IMediaDeleteChatSessionListener) ListenerManager.getInstance().removeListener(this.mKey);
-        if (iMediaDeleteChatSessionListener != null) {
-            iMediaDeleteChatSessionListener.onMediaDeleteChatSessionResult(i3, str);
+            IMediaDeleteChatSessionListener iMediaDeleteChatSessionListener = (IMediaDeleteChatSessionListener) ListenerManager.getInstance().removeListener(this.mKey);
+            if (iMediaDeleteChatSessionListener != null) {
+                iMediaDeleteChatSessionListener.onMediaDeleteChatSessionResult(i3, str);
+            }
         }
     }
 
@@ -126,6 +165,20 @@ public class IMMediaDeleteSessionRequest extends IMMediaBaseHttpRequest {
     }
 
     public IMMediaDeleteSessionRequest(Context context, long j, int i2, long j2, String str, long j3, String str2) {
+        Interceptable interceptable = $ic;
+        if (interceptable != null) {
+            InitContext newInitContext = TitanRuntime.newInitContext();
+            newInitContext.initArgs = r2;
+            Object[] objArr = {context, Long.valueOf(j), Integer.valueOf(i2), Long.valueOf(j2), str, Long.valueOf(j3), str2};
+            interceptable.invokeUnInit(65536, newInitContext);
+            int i3 = newInitContext.flag;
+            if ((i3 & 1) != 0) {
+                int i4 = i3 & 2;
+                newInitContext.thisArg = this;
+                interceptable.invokeInitBody(65536, newInitContext);
+                return;
+            }
+        }
         this.mContactorType = -1;
         this.mContactorPauid = -1L;
         this.mContext = context;
