+++ com/baidu/android/imsdk/task/TaskManager.java
@@ -1,7 +1,15 @@
 package com.baidu.android.imsdk.task;
 
 import android.content.Context;
+import com.baidu.android.imsdk.internal.Constants;
 import com.baidu.android.imsdk.utils.LogUtils;
+import com.baidu.titan.sdk.runtime.ClassClinitInterceptable;
+import com.baidu.titan.sdk.runtime.ClassClinitInterceptorStorage;
+import com.baidu.titan.sdk.runtime.FieldHolder;
+import com.baidu.titan.sdk.runtime.InitContext;
+import com.baidu.titan.sdk.runtime.InterceptResult;
+import com.baidu.titan.sdk.runtime.Interceptable;
+import com.baidu.titan.sdk.runtime.TitanRuntime;
 import java.util.concurrent.Callable;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
@@ -11,34 +19,78 @@ import java.util.concurrent.ThreadPoolExecutor;
 import java.util.concurrent.TimeUnit;
 /* loaded from: classes.dex */
 public class TaskManager {
+    public static /* synthetic */ Interceptable $ic = null;
     public static final int CORE_POOL_SIZE;
     public static final int CPU_COUNT;
     public static final int KEEP_ALIVE_SECONDS = 30;
     public static final int MAXIMUM_POOL_SIZE;
     public static final String TAG = "TaskManager";
     public static volatile TaskManager mInstance;
+    public transient /* synthetic */ FieldHolder $fh;
     public ThreadPoolExecutor service;
     public ExecutorService singleThreadService;
 
     /* loaded from: classes.dex */
     public static class Task implements Runnable {
+        public static /* synthetic */ Interceptable $ic;
+        public transient /* synthetic */ FieldHolder $fh;
         public String mAction;
         public String mJson;
 
         public Task() {
+            Interceptable interceptable = $ic;
+            if (interceptable != null) {
+                InitContext newInitContext = TitanRuntime.newInitContext();
+                interceptable.invokeUnInit(65536, newInitContext);
+                int i2 = newInitContext.flag;
+                if ((i2 & 1) != 0) {
+                    int i3 = i2 & 2;
+                    newInitContext.thisArg = this;
+                    interceptable.invokeInitBody(65536, newInitContext);
+                }
+            }
         }
 
         @Override // java.lang.Runnable
         public void run() {
+            Interceptable interceptable = $ic;
+            if (interceptable == null || interceptable.invokeV(1048576, this) == null) {
+            }
         }
 
         public Task(String str, String str2) {
+            Interceptable interceptable = $ic;
+            if (interceptable != null) {
+                InitContext newInitContext = TitanRuntime.newInitContext();
+                newInitContext.initArgs = r2;
+                Object[] objArr = {str, str2};
+                interceptable.invokeUnInit(65537, newInitContext);
+                int i2 = newInitContext.flag;
+                if ((i2 & 1) != 0) {
+                    int i3 = i2 & 2;
+                    newInitContext.thisArg = this;
+                    interceptable.invokeInitBody(65537, newInitContext);
+                    return;
+                }
+            }
             this.mAction = str;
             this.mJson = str2;
         }
     }
 
     static {
+        InterceptResult invokeClinit;
+        ClassClinitInterceptable classClinitInterceptable = ClassClinitInterceptorStorage.$ic;
+        if (classClinitInterceptable != null && (invokeClinit = classClinitInterceptable.invokeClinit(975634838, "Lcom/baidu/android/imsdk/task/TaskManager;")) != null) {
+            Interceptable interceptable = invokeClinit.interceptor;
+            if (interceptable != null) {
+                $ic = interceptable;
+            }
+            if ((invokeClinit.flags & 1) != 0) {
+                classClinitInterceptable.invokePostClinit(975634838, "Lcom/baidu/android/imsdk/task/TaskManager;");
+                return;
+            }
+        }
         int availableProcessors = Runtime.getRuntime().availableProcessors();
         CPU_COUNT = availableProcessors;
         CORE_POOL_SIZE = Math.max(4, Math.min(availableProcessors - 1, 4));
@@ -46,6 +98,18 @@ public class TaskManager {
     }
 
     public TaskManager() {
+        Interceptable interceptable = $ic;
+        if (interceptable != null) {
+            InitContext newInitContext = TitanRuntime.newInitContext();
+            interceptable.invokeUnInit(65537, newInitContext);
+            int i2 = newInitContext.flag;
+            if ((i2 & 1) != 0) {
+                int i3 = i2 & 2;
+                newInitContext.thisArg = this;
+                interceptable.invokeInitBody(65537, newInitContext);
+                return;
+            }
+        }
         this.service = null;
         this.singleThreadService = null;
         ThreadPoolExecutor.DiscardOldestPolicy discardOldestPolicy = new ThreadPoolExecutor.DiscardOldestPolicy();
@@ -56,39 +120,57 @@ public class TaskManager {
     }
 
     public static TaskManager getInstance(Context context) {
-        if (mInstance == null) {
-            synchronized (TaskManager.class) {
-                if (mInstance == null) {
-                    mInstance = new TaskManager();
+        InterceptResult invokeL;
+        Interceptable interceptable = $ic;
+        if (interceptable == null || (invokeL = interceptable.invokeL(65538, null, context)) == null) {
+            if (mInstance == null) {
+                synchronized (TaskManager.class) {
+                    if (mInstance == null) {
+                        mInstance = new TaskManager();
+                    }
                 }
             }
+            return mInstance;
         }
-        return mInstance;
+        return (TaskManager) invokeL.objValue;
     }
 
     public void clearTask() {
-        this.service.purge();
+        Interceptable interceptable = $ic;
+        if (interceptable == null || interceptable.invokeV(1048576, this) == null) {
+            this.service.purge();
+        }
     }
 
     public <T> Future<T> submitForLocalCallable(Callable<T> callable) {
-        return this.singleThreadService.submit(callable);
+        InterceptResult invokeL;
+        Interceptable interceptable = $ic;
+        return (interceptable == null || (invokeL = interceptable.invokeL(Constants.METHOD_GET_CONTACTER_INFO_FOR_SESSION, this, callable)) == null) ? this.singleThreadService.submit(callable) : (Future) invokeL.objValue;
     }
 
     public void submitForLocalOperation(Runnable runnable) {
-        try {
-            this.singleThreadService.submit(runnable);
-        } catch (Exception e2) {
-            LogUtils.e("TaskManager", "Exception ", e2);
+        Interceptable interceptable = $ic;
+        if (interceptable == null || interceptable.invokeL(Constants.METHOD_SEND_USER_MSG, this, runnable) == null) {
+            try {
+                this.singleThreadService.submit(runnable);
+            } catch (Exception e2) {
+                LogUtils.e("TaskManager", "Exception ", e2);
+            }
         }
     }
 
     public boolean submitForNetWork(Runnable runnable) {
-        try {
-            this.service.submit(runnable);
-            return true;
-        } catch (Throwable th) {
-            LogUtils.e("TaskManager", "Exception ", th);
-            return false;
+        InterceptResult invokeL;
+        Interceptable interceptable = $ic;
+        if (interceptable == null || (invokeL = interceptable.invokeL(1048579, this, runnable)) == null) {
+            try {
+                this.service.submit(runnable);
+                return true;
+            } catch (Throwable th) {
+                LogUtils.e("TaskManager", "Exception ", th);
+                return false;
+            }
         }
+        return invokeL.booleanValue;
     }
 }
