+++ com/baidu/mapsdkplatform/comapi/c.java
@@ -5,16 +5,22 @@ import android.content.Context;
 import android.util.Log;
 import com.baidu.mapapi.JNIInitializer;
 import com.baidu.mapapi.common.EnvironmentUtilities;
+import com.baidu.titan.sdk.runtime.FieldHolder;
+import com.baidu.titan.sdk.runtime.InterceptResult;
+import com.baidu.titan.sdk.runtime.Interceptable;
 import java.io.File;
 import java.io.IOException;
-/* loaded from: classes2.dex */
+/* loaded from: classes3.dex */
 public class c {
+    public static /* synthetic */ Interceptable $ic;
 
     /* renamed from: a  reason: collision with root package name */
-    public static boolean f7569a;
+    public static boolean f7599a;
+    public transient /* synthetic */ FieldHolder $fh;
 
     public static void a(Context context, boolean z, String str, String str2, String str3) {
-        if (f7569a) {
+        Interceptable interceptable = $ic;
+        if (!(interceptable == null || interceptable.invokeCommon(65536, null, new Object[]{context, Boolean.valueOf(z), str, str2, str3}) == null) || f7599a) {
             return;
         }
         if (context == null) {
@@ -34,27 +40,32 @@ public class c {
         }
         EnvironmentUtilities.initAppDirectory(context);
         com.baidu.mapsdkplatform.comapi.b.a.c.a().a(context);
-        f7569a = true;
+        f7599a = true;
     }
 
     public static boolean a(String str) {
-        if (str == null || str.isEmpty()) {
-            return false;
-        }
-        try {
-            File file = new File(str + "/check.0");
-            if (file.exists()) {
-                file.delete();
+        InterceptResult invokeL;
+        Interceptable interceptable = $ic;
+        if (interceptable == null || (invokeL = interceptable.invokeL(65537, null, str)) == null) {
+            if (str == null || str.isEmpty()) {
+                return false;
             }
-            file.createNewFile();
-            if (file.exists()) {
-                file.delete();
+            try {
+                File file = new File(str + "/check.0");
+                if (file.exists()) {
+                    file.delete();
+                }
+                file.createNewFile();
+                if (file.exists()) {
+                    file.delete();
+                    return true;
+                }
                 return true;
+            } catch (IOException e2) {
+                Log.e("SDKInitializer", "SDCard cache path invalid", e2);
+                throw new IllegalArgumentException("BDMapSDKException: Provided sdcard cache path invalid can not used.");
             }
-            return true;
-        } catch (IOException e2) {
-            Log.e("SDKInitializer", "SDCard cache path invalid", e2);
-            throw new IllegalArgumentException("BDMapSDKException: Provided sdcard cache path invalid can not used.");
         }
+        return invokeL.booleanValue;
     }
 }
